feat(knowledge): KnowledgeHealth shell — overview strip + analytics/verification/tools tabs (#11558)
