+++ autobot-frontend/src/components/knowledge/KnowledgeHealth.stories.ts
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 import type { Meta } from '@storybook/vue3'
 import type { StoryObj } from '@storybook/vue3'
-import KnowledgeMaintenance from './KnowledgeMaintenance.vue'
+import KnowledgeHealth from './KnowledgeHealth.vue'
 
 const meta = {
-  title: 'Components/Knowledge/KnowledgeMaintenance',
-  component: KnowledgeMaintenance,
+  title: 'Components/Knowledge/KnowledgeHealth',
+  component: KnowledgeHealth,
   tags: ['autodocs'],
-} as Meta<typeof KnowledgeMaintenance>
+} as Meta<typeof KnowledgeHealth>
 
 export default meta
 // #7273: relaxed to StoryObj<Record<string, unknown>> for render-only stories that don't match component props
