+++ autobot-frontend/src/components/knowledge/index.ts
@@ -30,7 +30,7 @@ export { default as KnowledgeGraph } from './KnowledgeGraph.vue'
 export { default as KnowledgeGraph3D } from './KnowledgeGraph3D.vue'
 export { default as KnowledgeGraphView } from './KnowledgeGraphView.vue'
 export { default as KnowledgeMainCategories } from './KnowledgeMainCategories.vue'
-export { default as KnowledgeMaintenance } from './KnowledgeMaintenance.vue'
+export { default as KnowledgeHealth } from './KnowledgeHealth.vue'
 export { default as KnowledgeManager } from './KnowledgeManager.vue'
 export { default as KnowledgePersistenceDialog } from './KnowledgePersistenceDialog.vue'
 export { default as KnowledgePromptEditor } from './KnowledgePromptEditor.vue'
